@@ -13,7 +13,6 @@
 
 package org.eclipse.jetty.client;
 
-import org.eclipse.jetty.client.api.Connection;
 import org.eclipse.jetty.util.Callback;
 import org.eclipse.jetty.util.Pool;
 import org.eclipse.jetty.util.annotation.ManagedAttribute;
@@ -22,71 +21,32 @@ import org.eclipse.jetty.util.annotation.ManagedObject;
 @ManagedObject
 public class MultiplexConnectionPool extends AbstractConnectionPool
 {
-    public MultiplexConnectionPool(HttpDestination destination, int maxConnections, Callback requester, int maxMultiplex)
+    public MultiplexConnectionPool(HttpDestination destination, int maxConnections, Callback requester, int initialMaxMultiplex)
     {
-        this(destination, maxConnections, false, requester, maxMultiplex);
+        this(destination, Pool.StrategyType.FIRST, maxConnections, false, requester, initialMaxMultiplex);
     }
 
-    public MultiplexConnectionPool(HttpDestination destination, int maxConnections, boolean cache, Callback requester, int maxMultiplex)
+    protected MultiplexConnectionPool(HttpDestination destination, Pool.StrategyType strategy, int maxConnections, boolean cache, Callback requester, int initialMaxMultiplex)
     {
-        this(destination, Pool.StrategyType.FIRST, maxConnections, cache, requester, maxMultiplex);
-    }
-
-    public MultiplexConnectionPool(HttpDestination destination, Pool.StrategyType strategy, int maxConnections, boolean cache, Callback requester, int maxMultiplex)
-    {
-        super(destination, new Pool<>(strategy, maxConnections, cache)
+        super(destination, new Pool<>(strategy, maxConnections, cache, connection ->
         {
-            @Override
-            protected int getMaxUsageCount(Connection connection)
-            {
-                int maxUsage = (connection instanceof MaxUsable)
-                    ? ((MaxUsable)connection).getMaxUsageCount()
-                    : super.getMaxUsageCount(connection);
-                return maxUsage > 0 ? maxUsage : -1;
-            }
-
-            @Override
-            protected int getMaxMultiplex(Connection connection)
-            {
-                int multiplex = (connection instanceof Multiplexable)
-                    ? ((Multiplexable)connection).getMaxMultiplex()
-                    : super.getMaxMultiplex(connection);
-                return multiplex > 0 ? multiplex : 1;
-            }
-        }, requester);
-        setMaxMultiplex(maxMultiplex);
-    }
-
-    @Deprecated
-    public MultiplexConnectionPool(HttpDestination destination, Pool<Connection> pool, Callback requester, int maxMultiplex)
-    {
-        super(destination, pool, requester);
-        setMaxMultiplex(maxMultiplex);
+            int maxMultiplex = initialMaxMultiplex;
+            if (connection instanceof MaxMultiplexable maxMultiplexable)
+                maxMultiplex = maxMultiplexable.getMaxMultiplex();
+            return maxMultiplex;
+        }), requester, initialMaxMultiplex);
     }
 
     @Override
-    @ManagedAttribute(value = "The multiplexing factor of connections")
-    public int getMaxMultiplex()
+    @ManagedAttribute(value = "The initial multiplexing factor of connections")
+    public int getInitialMaxMultiplex()
     {
-        return super.getMaxMultiplex();
+        return super.getInitialMaxMultiplex();
     }
 
     @Override
-    public void setMaxMultiplex(int maxMultiplex)
+    public void setInitialMaxMultiplex(int initialMaxMultiplex)
     {
-        super.setMaxMultiplex(maxMultiplex);
-    }
-
-    @Override
-    @ManagedAttribute(value = "The maximum amount of times a connection is used before it gets closed")
-    public int getMaxUsageCount()
-    {
-        return super.getMaxUsageCount();
-    }
-
-    @Override
-    public void setMaxUsageCount(int maxUsageCount)
-    {
-        super.setMaxUsageCount(maxUsageCount);
+        super.setInitialMaxMultiplex(initialMaxMultiplex);
     }
 }
