@@ -31,7 +31,7 @@ import org.eclipse.jetty.http3.client.internal.HTTP3SessionClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class HttpConnectionOverHTTP3 extends HttpConnection implements ConnectionPool.Multiplexable, ConnectionPool.MaxUsable
+public class HttpConnectionOverHTTP3 extends HttpConnection implements ConnectionPool.MaxMultiplexable, ConnectionPool.MaxUsable
 {
     private static final Logger LOG = LoggerFactory.getLogger(HttpConnectionOverHTTP3.class);
 
@@ -55,11 +55,11 @@ public class HttpConnectionOverHTTP3 extends HttpConnection implements Connectio
     {
         // As weird as this is, RFC 9000 specifies a *cumulative* number
         // for the number of streams that can be opened in a connection.
-        return getMaxUsageCount();
+        return getMaxUsage();
     }
 
     @Override
-    public int getMaxUsageCount()
+    public int getMaxUsage()
     {
         return session.getMaxLocalStreams();
     }
