@@ -17,6 +17,7 @@ import java.io.IOException;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
+
 import jakarta.servlet.AsyncContext;
 import jakarta.servlet.ReadListener;
 import jakarta.servlet.ServletException;
@@ -27,11 +28,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.eclipse.jetty.client.api.ContentResponse;
 import org.eclipse.jetty.client.util.StringRequestContent;
-import org.eclipse.jetty.ee9.http.client.AbstractTest;
-import org.eclipse.jetty.ee9.http.client.EmptyServerHandler;
-import org.eclipse.jetty.ee9.http.client.Transport;
-import org.eclipse.jetty.ee9.http.client.TransportProvider;
-import org.eclipse.jetty.ee9.http.client.TransportScenario;
 import org.eclipse.jetty.http.HttpMethod;
 import org.eclipse.jetty.http.HttpStatus;
 import org.eclipse.jetty.server.Request;
